process/scaleway/process_eex_data/requirements.txt
openpyxl
psycopg2
pandas
numpy
boto3
python-dateutil